+++ src/Network/Network/ChangeLog.md
@@ -31,6 +31,7 @@
     - `New-AzPublicIpAddress`
 * Fixed bug where HubRoutingPreference didn't show up when running 'Get-AzRouteServer'
 * Updated `New-AzVirtualNetworkGateway` to remove validation for `ExtendedLocation` parameter
+* Add support for new Application Gateway SKU type, Basic SKU
 
 ## Version 6.0.0
 * Added new cmdlets for RouteMap child resource of VirtualHub.